refactor(user): replace getStoredItem with useUser hook for user state management (#194)
* feat(user): add user context and provider for managing authentication

Introduce a new user context and provider to centralize authentication logic. This includes handling login, logout, and user profile management. The `UserProvider` is integrated into the main app component, and the `useUser` hook is utilized in the Discord connection page to manage user state and authentication. This refactor improves code maintainability and reduces redundancy in authentication logic.

* refactor(user): replace token checks with isConnected from user store

Refactored multiple components to use the `isConnected` state from the user store instead of directly checking the token. This improves code maintainability and centralizes the authentication state management.

* fix(userStore): set discordId in login function to ensure state consistency

The login function now sets the discordId in the state to ensure consistency between the stored item and the state. This prevents potential discrepancies when accessing the discordId during the session.

* refactor(user): replace getStoredItem with useUser hook for user state management

Refactored multiple components to use the `useUser` hook instead of directly calling `getStoredItem` for user-related data. This centralizes user state management and improves code maintainability. Additionally, removed the unused `error` state from the user store to simplify the logic.

* refactor(DiscordConnection): remove unused import and optimize conditionals

Remove the unused `getStoredItem` import and simplify conditional checks using optional chaining. Also, add `isConnected` to the dependency array of `useMemo` to ensure proper reactivity.